Updated Code + creation of backend
a backend was made, a register and login feature was created. Manual input for admin credentials is required so that we can access the admin webAPP. A revision of forms is needed so that when creating an order it shows a overview of what it is before sending you to checkout. The forms also needs revision so that it has a cleaner feel and look. We need to focus on trying to connect the requests to the order table inside the admin orders table AFTER proceeding to checkout.